requirements.txt
aiogram
python-dotenv
pillow
google-genai
